bi_generate_product_ean13/requirements.txt
python-barcode==0.9.0
